+++ tests/conftest.py
@@ -178,6 +178,7 @@ def demo_messages_from_data_source_with_test_streams(
 
 @pytest.fixture
 def general_data() -> List[dict]:
+    """21 event."""
     data = [
         {
             "batchId": None,
@@ -356,6 +357,165 @@ def general_data() -> List[dict]:
             "isBatched": False,
             "eventType": "Checkpoint for session",
             "parentEventId": "88a3ee80-d1b4-11eb-b0fb-199708acc7bc",
+        },
+    ]
+    return data
+
+
+@pytest.fixture
+def detached_data() -> List[dict]:
+    data = [
+        {
+            "batchId": None,
+            "eventId": "84db48fc-d1b4-11eb-b0fb-199708acc7bc",
+            "eventName": "[TS_1]Aggressive IOC vs two orders: second order's price is " "lower than first",
+            "eventType": "",
+            "isBatched": False,
+            "parentEventId": None,
+        },
+        {
+            "batchId": None,
+            "eventId": "88a3ee80-d1b4-11eb-b0fb-199708acc7bc",
+            "eventName": "Case[TC_1.1]: Trader DEMO-CONN1 vs trader DEMO-CONN2 for " "instrument INSTR1",
+            "eventType": "",
+            "isBatched": False,
+            "parentEventId": "84db48fc-d1b4-11eb-b0fb-199708acc7bc",
+        },
+        {
+            "batchId": None,
+            "eventId": "8bc787fe-d1b4-11eb-bae5-57b0c4472880",
+            "eventName": 'placeOrderFIX demo-conn1 - STEP1: Trader "DEMO-CONN1" sends '
+            "request to create passive Order.",
+            "eventType": "placeOrderFIX",
+            "isBatched": False,
+            "parentEventId": "88a3ee80-d1b4-11eb-b0fb-199708acc7bc",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint",
+            "eventType": "Checkpoint",
+            "isBatched": True,
+            "parentEventId": "8bc787fe-d1b4-11eb-bae5-57b0c4472880",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114a4-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'th2-hand-demo' direction 'FIRST' "
+            "sequence '1623852603564709030'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114a5-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-conn1' direction 'SECOND' "
+            "sequence '1624005455622140289'",
+            "eventType": "Checkpoint for session",
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114a6-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-dc1' direction 'SECOND' " "sequence '1624005475721015014'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114a7-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-dc1' direction 'FIRST' " "sequence '1624005475720919499'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114a8-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-conn2' direction 'FIRST' "
+            "sequence '1624005448022245399'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114a9-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-conn2' direction 'SECOND' "
+            "sequence '1624005448022426113'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114aa-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-dc2' direction 'SECOND' " "sequence '1624005466840347015'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114ab-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-dc2' direction 'FIRST' " "sequence '1624005466840263372'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114ac-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-conn1' direction 'FIRST' "
+            "sequence '1624005455622011522'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c035903-d1b4-11eb-9278-591e568ad66e",
+        },
+        {
+            "batchId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4",
+            "eventId": "6e3be13f-cab7-4653-8cb9-6e74fd95ade4:8c1114ad-d1b4-11eb-9278-591e568ad66e",
+            "eventName": "Checkpoint for session alias 'demo-log' direction 'FIRST' " "sequence '1624029363623063053'",
+            "eventType": "Checkpoint for session",
+            "isBatched": True,
+            "parentEventId": None,
+        },
+        {
+            "batchId": None,
+            "eventId": "8c3fec4f-d1b4-11eb-bae5-57b0c4472880",
+            "eventName": "Send 'NewOrderSingle' message to connectivity",
+            "eventType": "Outgoing message",
+            "isBatched": False,
+            "parentEventId": "8bc787fe-d1b4-11eb-bae5-57b0c4472880",
+        },
+        {
+            "batchId": None,
+            "eventId": "8c44806c-d1b4-11eb-8e55-d3a76285d588",
+            "eventName": "Send 'NewOrderSingle' message",
+            "eventType": "Outgoing message",
+            "isBatched": False,
+            "parentEventId": "8bc787fe-d1b4-11eb-bae5-57b0c4472880",
+        },
+        {
+            "batchId": "654c2724-5202-460b-8e6c-a7ee9fb02ddf",
+            "eventId": "654c2724-5202-460b-8e6c-a7ee9fb02ddf:8ca20288-d1b4-11eb-986f-1e8d42132387",
+            "eventName": "Remove 'NewOrderSingle' "
+            "id='demo-conn1:SECOND:1624005455622135205' "
+            "Hash='7009491514226292581' Group='NOS_CONN' "
+            "Hash['SecondaryClOrdID': 11111, 'SecurityID': INSTR1]",
+            "isBatched": True,
+            "eventType": "",
+            "parentEventId": None,
+        },
+        {
+            "batchId": None,
+            "eventId": "8ceb47f6-d1b4-11eb-a9ed-ffb57363e013",
+            "eventName": "Send 'ExecutionReport' message",
+            "isBatched": False,
+            "eventType": "Send message",
+            "parentEventId": "845d70d2-9c68-11eb-8598-691ebd7f413d",
         },
     ]
     return data